+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to the "leadwerks" extension will be documented in this file.
 
+
+### 0.6.7
+Paramters are added to the funtions. Press tab when inserting a snippet to jump between paramters. 
+Description contains explanation of the parameters
+
+### 0.6.x
+First generation of Leadwerks API snippets based on the Toc.xml. 
+
 ### 0.4.2
 Added 'lescripts' snippet that adds all default leadwerks entity script functions.
 Matched the minor version with the package.json level